Add simple trait id to the definition vampire change
Sets a property on the SimpleTrait itself so that after saving it we can add the simple_trait_id to the define VampireChange. Will need to continue tracking with the assumption that this could be blank since old data won't have this and a failure in afterSave doesn't stop anything on the client side.